git-svn-id: http://pcsx2.googlecode.com/svn/branches/GregMiscellaneous@3679 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zzogl-pg/opengl/ZZoglFlush.cpp
@@ -118,7 +118,6 @@ void Draw(const VB& curvb)
 
 //------------------ variables
 
-//extern bool g_bIsLost;
 extern int g_nDepthBias;
 extern float g_fBlockMult;
 bool g_bUpdateStencil = 1;
@@ -377,7 +376,7 @@ inline void FlushUpdateEffect()
 // Check, maybe we cold skip flush
 inline bool IsFlushNoNeed(VB& curvb, const pixTest& curtest)
 {
-	if (curvb.nCount == 0 || (curtest.zte && curtest.ztst == 0) /*|| g_bIsLost*/ || IsBadFrame(curvb))
+	if (curvb.nCount == 0 || (curtest.zte && curtest.ztst == 0) || IsBadFrame(curvb))
 	{
 		curvb.nCount = 0;
 		return true;
@@ -2033,7 +2032,7 @@ void ZeroGS::SetTexClamping(int context, FRAGMENTSHADER* pfragment)
 
 		case 3: // region rep x
 			v.x = 0.9999f;
-			v.z = fw;
+			v.z = (float)fw;
 			v2.x = (float)GPU_TEXMASKWIDTH / fw;
 			v2.z = pclamp->maxu / fw;
 			int correctMinu = pclamp->minu & (~pclamp->maxu);		// (A && B) || C == (A && (B && !C)) + C
@@ -2068,7 +2067,7 @@ void ZeroGS::SetTexClamping(int context, FRAGMENTSHADER* pfragment)
 
 		case 3: // region rep y
 			v.y = 0.9999f;
-			v.w = fh;
+			v.w = (float)fh;
 			v2.y = (float)GPU_TEXMASKWIDTH / fh;
 			v2.w = pclamp->maxv / fh;
 			int correctMinv = pclamp->minv & (~pclamp->maxv);		// (A && B) || C == (A && (B && !C)) + C
